MixHighTech - AdminCoupons OK

+++ Projet1/src/main/java/fr/afcepf/al29/groupem/business/api/CouponBusApi.java
@@ -27,8 +27,8 @@ public interface CouponBusApi {
     List<Coupon> getCouponsEndingAfter(Date date);
     List<Coupon> getCouponsEndingOn(Date date);
     
-    List<Coupon> getCouponsByRebateLesserThan(int rebate);
-    List<Coupon> getCouponsByRebateGreaterThan(int rebate);
-    List<Coupon> getCouponsByRebateEquals(int rebate);
+    List<Coupon> getCouponsByRebateLesserThan(float rebate);
+    List<Coupon> getCouponsByRebateGreaterThan(float rebate);
+    List<Coupon> getCouponsByRebateEquals(float rebate);
 
 }
